+++ default.nix
@@ -14,8 +14,9 @@ pkgs.stdenv.mkDerivation rec {
     pkgs.meson pkgs.ninja pkgs.pkgconfig pkgs.asciidoc pkgs.libxslt.bin
     pkgs.docbook_xml_dtd_45 pkgs.docbook_xsl pkgs.libxml2.bin pkgs.docbook5
     pkgs.python3Packages.pytest pkgs.python3Packages.pytest-timeout
+    pkgs.systemd
   ];
-  buildInputs = [ pkgs.libyamlcpp pkgs.systemd ];
+  buildInputs = [ pkgs.libyamlcpp ];
 
   doCheck = true;